+++ Projet/printf/specifiersxo_lengthjz.c
@@ -6,7 +6,7 @@
 /*   By: qmoricea <marvin@le-101.fr>                +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/12/03 10:17:19 by qmoricea     #+#   ##    ##    #+#       */
-/*   Updated: 2018/12/03 10:17:50 by qmoricea    ###    #+. /#+    ###.fr     */
+/*   Updated: 2018/12/18 09:47:21 by qmoricea    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -33,15 +33,15 @@ int		ft_printf_specz(const char *format, va_list ap, int i)
 
 	all = 0;
 	if (format[i] == 'z' && (format[i + 1] == 'd' || format[i + 1] == 'i'))
-		all += ft_printf_nbr((size_t)va_arg(ap, intmax_t));
+		all += ft_printf_nbr(va_arg(ap, size_t));
 	else if (format[i] == 'z' && format[i + 1] == 'u')
-		all += ft_printf_unbr((size_t)va_arg(ap, uintmax_t));
+		all += ft_printf_unbr(va_arg(ap, size_t));
 	else if (format[i] == 'z' && format[i + 1] == 'o')
-		all += ft_printf_octal((size_t)va_arg(ap, intmax_t));
+		all += ft_printf_octal(va_arg(ap, size_t));
 	else if (format[i] == 'z' && format[i + 1] == 'x')
-		all += ft_printf_x((size_t)va_arg(ap, intmax_t));
+		all += ft_printf_x(va_arg(ap, size_t));
 	else if (format[i] == 'z' && format[i + 1] == 'X')
-		all += ft_printf_xm((size_t)va_arg(ap, intmax_t));
+		all += ft_printf_xm(va_arg(ap, size_t));
 	return (all);
 }
 
@@ -51,15 +51,15 @@ int		ft_printf_specjz(const char *format, va_list ap, int i)
 
 	all = 0;
 	if (format[i] == 'j' && (format[i + 1] == 'd' || format[i + 1] == 'i'))
-		all += ft_printf_nbr(va_arg(ap, intmax_t));
+		all += ft_printf_nbr(va_arg(ap, uintmax_t));
 	else if (format[i] == 'j' && format[i + 1] == 'u')
-		all += ft_printf_unbr((uintmax_t)va_arg(ap, intmax_t));
+		all += ft_printf_unbr(va_arg(ap, uintmax_t));
 	else if (format[i] == 'j' && format[i + 1] == 'o')
-		all += ft_printf_octal((uintmax_t)va_arg(ap, intmax_t));
+		all += ft_printf_octal(va_arg(ap, uintmax_t));
 	else if (format[i] == 'j' && format[i + 1] == 'x')
-		all += ft_printf_x((uintmax_t)va_arg(ap, intmax_t));
+		all += ft_printf_x(va_arg(ap, uintmax_t));
 	else if (format[i] == 'j' && format[i + 1] == 'X')
-		all += ft_printf_xm((uintmax_t)va_arg(ap, intmax_t));
+		all += ft_printf_xm(va_arg(ap, uintmax_t));
 	else if (format[i] == 'z')
 		all += ft_printf_specz(format, ap, i);
 	return (all);
